refactor(routing): rename target resolver package
Amp-Thread-ID: https://ampcode.com/threads/T-019eb357-321a-7760-b4fe-e3c606cba251
