+++ pkg/api/config.go
@@ -14,9 +14,9 @@ import (
 	"time"
 
 	"github.com/block/schemabot/pkg/pendingdrops"
+	"github.com/block/schemabot/pkg/routing"
 	"github.com/block/schemabot/pkg/secrets"
 	"github.com/block/schemabot/pkg/storage"
-	"github.com/block/schemabot/pkg/targetresolver"
 	gomysql "github.com/go-sql-driver/mysql"
 	"gopkg.in/yaml.v3"
 )
@@ -845,13 +845,13 @@ func (c *ServerConfig) DatabaseEnvironments(database string) ([]string, error) {
 // For environments configured with a deployments map (multi-deployment), this
 // returns the single deployment when exactly one is configured and otherwise
 // errors. Callers that need the full set MUST use ResolveDatabaseTargets.
-func (c *ServerConfig) ResolveDatabaseTarget(database, environment string) (targetresolver.Target, error) {
+func (c *ServerConfig) ResolveDatabaseTarget(database, environment string) (routing.ExecutionTarget, error) {
 	targets, err := c.ResolveDatabaseTargets(database, environment)
 	if err != nil {
-		return targetresolver.Target{}, err
+		return routing.ExecutionTarget{}, err
 	}
 	if len(targets) != 1 {
-		return targetresolver.Target{}, fmt.Errorf("database %q environment %q resolves to %d deployments; use ResolveDatabaseTargets", database, environment, len(targets))
+		return routing.ExecutionTarget{}, fmt.Errorf("database %q environment %q resolves to %d deployments; use ResolveDatabaseTargets", database, environment, len(targets))
 	}
 	return targets[0], nil
 }
@@ -909,9 +909,9 @@ func validateDeploymentOrder(deployments map[string]DeploymentTarget, order []st
 	return nil
 }
 
-// ResolveTargets implements targetresolver.Resolver using this server's static
+// ResolveTargets implements routing.Resolver using this server's static
 // configuration.
-func (c *ServerConfig) ResolveTargets(_ context.Context, req targetresolver.Request) ([]targetresolver.Target, error) {
+func (c *ServerConfig) ResolveTargets(_ context.Context, req routing.Request) ([]routing.ExecutionTarget, error) {
 	return c.ResolveDatabaseTargets(req.Database, req.Environment)
 }
 
@@ -920,7 +920,7 @@ func (c *ServerConfig) ResolveTargets(_ context.Context, req targetresolver.Requ
 // configurations (scalar target/deployment or local DSN) return a one-element
 // slice; multi-deployment environments return one element per entry in the
 // deployments map, ordered deterministically by deployment key.
-func (c *ServerConfig) ResolveDatabaseTargets(database, environment string) ([]targetresolver.Target, error) {
+func (c *ServerConfig) ResolveDatabaseTargets(database, environment string) ([]routing.ExecutionTarget, error) {
 	if c == nil {
 		return nil, fmt.Errorf("server config is nil")
 	}
@@ -934,7 +934,7 @@ func (c *ServerConfig) ResolveDatabaseTargets(database, environment string) ([]t
 	}
 
 	if envConfig.HasLocalDSN() {
-		return []targetresolver.Target{{
+		return []routing.ExecutionTarget{{
 			DatabaseType: dbConfig.Type,
 			Deployment:   database,
 			Target:       database,
@@ -953,13 +953,13 @@ func (c *ServerConfig) ResolveDatabaseTargets(database, environment string) ([]t
 		if err != nil {
 			return nil, err
 		}
-		out := make([]targetresolver.Target, 0, len(deployments))
+		out := make([]routing.ExecutionTarget, 0, len(deployments))
 		for _, deployment := range deployments {
 			dt := envConfig.Deployments[deployment]
 			if dt.Target == "" {
 				return nil, fmt.Errorf("database %q environment %q deployment %q missing target", database, environment, deployment)
 			}
-			out = append(out, targetresolver.Target{
+			out = append(out, routing.ExecutionTarget{
 				DatabaseType: dbConfig.Type,
 				Deployment:   deployment,
 				Target:       dt.Target,
@@ -974,7 +974,7 @@ func (c *ServerConfig) ResolveDatabaseTargets(database, environment string) ([]t
 	if envConfig.Deployment == "" {
 		return nil, fmt.Errorf("database %q environment %q missing server-side deployment", database, environment)
 	}
-	return []targetresolver.Target{{
+	return []routing.ExecutionTarget{{
 		DatabaseType: dbConfig.Type,
 		Deployment:   envConfig.Deployment,
 		Target:       envConfig.Target,
